-fixed mercs complaining about equipment when it's really good -fixed BR not getting new stock for items with indexes > 1000 -reduced re-ordering times for awesome BR's
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@357 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Strategic/Strategic Merc Handler.cpp
@@ -689,10 +689,10 @@ BOOLEAN SoldierHasWorseEquipmentThanUsedTo( SOLDIERTYPE *pSoldier )
 {
 	INT32		cnt;
 	UINT16	usItem;
-	INT8		bBestArmour = -1;
-	INT8		bBestArmourIndex = -1;
-	INT8		bBestGun = -1;
-	INT8		bBestGunIndex = -1;
+	INT32		bBestArmour = -1;
+	INT32		bBestArmourIndex = -1;
+	INT32		bBestGun = -1;
+	INT32		bBestGunIndex = -1;
 
 
 	for ( cnt = 0; cnt < NUM_INV_SLOTS; cnt++ )
